Testing Some Basic Ideas for image detection in opencv
Been trying to get Canny edge detection to work, this will work for angled images or curved lines but not for a straight line. Also tested out Harris corner detection and Hough line transform, could not get either of these to work properly. For Harris corner detection converting to float32 with numpy did not seem to work so the corner detection algorithm would not work on the file. For Hough line transform a binary image is needed which can be created using cv2.threshold but this will not work on a none greyscaled image, could not get the image to properly convert to greyscale. I think if I get Canny edge detection working then I could use Hough line transform to find an equation for the lines and possibly the length.