Stops Files From Overwriting when saved
Added a function saveImg() that will look check the path where the image wants to be saved if a file already exists. If so the function will add a number to the end of the file so it doesn't overwrite anything. Will number files indefinately so it won't ever overwrite an image. The original plan was to have the plt.saveFig() function inside my defined function but this did not seem to be working, so all the function does is finds a free file name and returns it. Have left so excess code in that will be looked at to see if the saveFig() function can be put in. If this cannot be done then the function will be edited to remove all the excess code, it doesn't cause any errors just isn't needed.