Noticed that Canny Edge Detection Had not been using blurred image so redid tests from last commit
Tried the previous tests ran with the blurred image being analysed by Canny function too, it was an error not to include it last commit. The kernel was again changed [4] 5*5, [5] 10*10, [6] 100*100, [7] 200*200. These do agree that canny edge detection works better with a gradient as it starts to work with the greater blurring, although this amount of blurring really changes the image so will not produce reliable results.