Trying to show the image with OpenCV Another Error
Trying to show the Hough line transform image with cv2.imshow() instead of using matplotlib but it crashes the ipython console. The image produce by the Hough transform is not able to show, for now I am going to ignore Hough line transform and concentrate on Harris corner detection because that seems the most promising.
Also the import of image from matplotlib was removed as it is no longer used.