Testing different variables for Harris Corner detection
Swapping back to using the grayscale image after removing the blurring function. Then editing the blockSize and kSize variables for the Harris corner detection method, this highlighted the corners properly on the image, but also I went over 10 copies of an image. This meant that the image number was in double figures and I hadn't thought to account for double figures in the saveImg() function so that has been edited to work with multiple digit numbers. This new algorithm has been tested up to filenames up to 105 in the ipython console.