Testing Harris corner detection Finding more accurate corners
Using cornerSubPix() to try and find the corners more accurately but does not seem to be correctly finding the corners, it only returns 3 sets of coordinates when there should be 4 corners found. The Harris corner detection image shows four corners in the right points but the connectedComponentsWithStats() function does no find the correct coordinates for these corners.