Testing the code with different image files
Testing it with some other images to see if it works any better, still showing the corners in the correct place when using cornerHarris() but will not find the correct centroid positions. Not sure why it only gives me 3 sets of coordinates and why they aren't in the correct place. Will have to try a method analyzing the numpy array to find the coloured points.