Creating a binary image of the corners
Created a binary image using threshold function so that corner pixels are white and everything else is black, then using numpy.nonzero() function to find the coordinates of these white pixels. However the corner detection algorithm is finding too large an area so there are too many coordinates for each corner. Need to find a way to find the centroid of the groups of pixels, opencv has this function but it isn't working correctly. Which is why cornerSubPix() isn't working. Not sure how to do this with numpy?