Plotting Histograms and Saving Harris Corner Image and SubPixel Image
Removed the plotting of the subPixel data because it returns an array of corner positions so the array is being printed instead of plotted as a graph. Have plotted these positions on the original image and saved it so I can see where it thinks corners are. Also saved an image of the Harris Corner detection to see where that function thinks the corners are. Printed total number of data points in original image histogram to check it is correct, getting 3 times the number of pixels because there are 3 values for each pixel for an rgb jpeg image.