Finding the midpoint of each short edge
Find the midpoint of any corners considered close enough to be either corner on the short edge of the fibre, then returning an array of all the midpoints. Giving you an array with the endpoint of every fibre, finding the length of the single fibre using the distance between the two end points. Tested with original verticle fibre then rotated at 90deg, 45deg, -45deg, 30deg and 60deg relative to original. For each image it found the length to be 500pixels when rounded to 1pixel. For each test it also only found the four corners expected no false corners were detected.