Function to generalise the measurement of length
Function that will measure the distance between every point in the array of the fibre endpoints. Checks for zero distance and ignores because this is measurement between the same position. Will everntually check the lines are joined before finding the distance, but that code has not been written yet.  Still only tested on single fibre image.