Testing with multiple fibres Also Removed Hough line transform (unused)
Created images with multiple fibres (copies of original) at 4 different angles, 0, 45, 90 and 135. For four fibres it finds only the corners for the top two fibres, the horizontal and vertical ones. Does not find the centroids for the ones at an angle, but does find them with the first Harris function.
Adding some print statements to print the number of corners, endpoints and lengths found, and it seems to only be finding the horizontal and vertical fibres for each case, but the corners still show up on the harris images.
Tested again with single fibre at an angle and it works fine.