Debugging Large 2 Fibre Image
Checking the arrays in the variable explorer. Can see the corner pos in corners array, the array the Harris image is produced from but cannot see anything in the cornersThres image. Therefore threshold() is causing the problem.

Change threshold to 0.8*max and correct corners are found.

Final test with 64 test fibres (25,500 fibre).jpg to see if its fixed, found only half the corners. Added line to save cornersThres array for easier debugging.
Trying with thresholding on 0.75*max. Still not low enough trying 0.7.
0.7*max finds the correct number of endpoints and corners and the corners are in the correct positions.