checkLine() function written
The function to check if two points are connected by a line of pixels has been written. It will find an equation for the line between the two points then loop through each position on the line and check that pixel is black. The checkBlack() function will check if the pixel is black (part of a fibre so maybe gray) but has not been written yet. It may also need to check surrounding pixels if they are off by one pixel. However this may need to be included as some sort of error.