Writing the checkBlack() function
Writing the function to check if all the pixels between two edge are black and therefore are part of the same fibre. Works with the single vertical fibre and works with two of the fibres in the four fibre image "4 test fibres (25,500 fibre).jpg" specifically the horizontal and vertical fibres. Currently not working with the fibres at an angle.