Fixing error with "45degree fibre test 2 (25,500 fibre).jpg"
Would not find the lengths for the fibres because it did not think they were connected this was due to blurring at the edge of the lines so the pixel can now be more gray in colour and still will count. Changing this allowed for one length to be found.
The other fibre would not find an equation for the line and this was due to rounding errors causing the checking of the yIntercept equation not to be corrected, this was solved by rounding the final values before checking if they are equal.