Tested on 64 fibres
Ran the program on an image with 64 fake fibres in it a different angles and it found the correct lengths for all of the fibres. However it took ~1hr 45mins, so more efficient methods need to be done. One possible one is checking the fibre every 10pixels is black instead of every single pixel.