Organising code
Remove all the text files from the code directory and put them in a new directory called "Test Files". Split the code into 5 separate modules the main one controlling all other functions, a module for controlling inputs, a module for finding the corners and edges, a module for finding the fibre lengths and finally and module for data analysis and producing the graphs. Currently not functioning as the functions need to be sorted in the modules and then called in the main file.