Editing the modules to work again
Adding some code to the main python file so that all the modules are imported and all the functions are ran from that file. Also added more constants at the top of the file so they can be changed easier, instead of having them within other functions. Also edited the saveImg() function to add a date/time the program started so all saved files could have the same name.
Changed the name of input.py to inputs.py because in conflicted with a used word in python.
Edited corners and lengths modules to improve the doc strings on the functions and added in an argument for the image array to the functions that needed to check against the original array.