Reducing number of lengths checked
Making the code more efficient by reducing the number of lengths that need to be checked, this is done by having a boolean numpy array is the same length as the coordinate array and refers to each coordinate. Once two coordinates have been joined they are both marked as True because they only connect to one another, this stops them from being checked with other coordinates and reduces the lengths needed to be checked.

Also changed the print outs so that it isn't number of lengths checked it is instead number of coordinates found.