Generating random fibre image
Added the function generateImage() that randomly generates an array containing some fibres, needs to be edited to it keeps track of positions and lengths. When testing the program would find some of the fibres but not all it struggled with crossing fibres because it found the corners at the positions where they crossed.

Tried to add in some code to draw lines on the fibres it had found however have yet to get it working, not sure of the issue.