Adding in ability to run mulitple images
Added in some while loops to main.py in order to allow multiple images to be analysed without input from a user. Also changed the input so the user could request a number of random images to be generated and analysed just by typing "Random 4" for example.

Edited inputs.generateImage() so that the maximum length is independant of array size, also converted the array to 8 bit before returning it as openCV was having some issues with it.

Created a drawFound() function in lengths.py and moved over the currently not working algorithm for colouring in found fibres.

Tested the program analysing muliple random images.