Testing and Fixing lengths.drawFound()
Writing some additional code in lengths so that that file can be ran on its own to test drawFound(), found errors in converting the fibreLengths array into int. This has been solved by using numpy to convert it to int, then rearranging the coordinates order so the lines are drawn in the correct position. This function now works correctly next it needs to be called from the main file, and the excess code in lengths removed.