Edited generateImage() and Tested
Changed generateImage() so that all the corners generated are rounded to the nearest int and also changed the lengths array so that it is a 1D array instead of a 2D one it was before due to how it was created in trig().
Tested with 30 random images and it found the correct corners and endpoints for all images and the correct number of lengths for all but one image. "Generated Random Image[2018-02-10_17-29-23] (Random Image 17)" it did not find one of the lengths but had found all the correct corners and endpoints.
One possible reason could be that the endpoints were a pixel off the fibre so it was white and therefore not considered part of a fibre, checkLine() function will need to be edited to check the surrounding pixels.