Drawing Midpoint Position on Image
Removed the saving of the image with the corners drawn on from corners.findCorners() and instead returned the imageArray with the corners added so it could be saved later.  In main.py added the midpoint positions to the array so there is the image array with the centroids, subpixel corners and the midpoints drawn on. Then saved this array as an image.