Fixing Analysing Crossing fibres
Edited corners.findCorners() so that it checks if the centroid position is part of a fibre and removes it if not, this removes the corners at fibre crossing points as the centroid point is alway on the inside of a corner. [2018-02-18_17-32-54]

Edited corners.averageEdges() so that it checks if the midpoint found is part of a fibre, checks all surrounding pixels. [2018-02-18_17-55-40]

Edited lengths.checkLine() in the if statement the or meant that it would carry on if only one of the two midpoints being checked were correct, changed to an and which seemed to fix that. [2018-02-18_18-09-36]