Testing on Random Image
When testing on a random image [2018-02-18_18-18-44]  debugging was turned on to see why some corners were not being found and it showed them being removed due to thresholding. So the tresholding level was reduced.

Reducing this thresholding caused the corners to be found correctly on the two fibres they were missing on  [2018-02-18_18-28-36], two fibres can still not be found but that is due to one or more of the corners being inside another fibre. It is likely there is no simple way of finding the length of these fibres because not all corners can be found.

Edited checking for random in the input so filenames could contain the word "Random" as long as it wasn't at the start.