Reverting code and testing
Removed some of  the edits made in commit b388da6e16fc81d899d1aa3b6340eb45f510a5fc to test the code after just the edits for checking the centroid positions are made. Removed checking midpoint is part of a fibre and removed checking the full line instead of half.

Tested and added data to manualGraphs as well as templates for the next sets of data.