PREVIOUS COMMIT INCORRECT
In commit 174e0a275e882cb89f904c9e8d3893a99a009c87 this was stated. "Removed some of  the edits made in commit b388da6 to test the code after just the edits for checking the centroid positions are made. Removed checking midpoint is part of a fibre and removed checking the full line instead of half.

Tested and added data to manualGraphs as well as templates for the next sets of data."

However actually the removed edits were the centroids position checks and checking the full line. The edit that was kept was the checking the midpoint was part of a fibre. The graph title has been updated to state this.