Editing Midpoint Finding
Added saving the lengths distributions to graphing.

Edited corners.averageEdges() so that any corners that a midpoint could not be found for will be added to the midpoints array. This means that if only one corner at the edge of a fibre has been found that will be used as the endpoint for that fibre when calculating the length, thus allowing for more fibres to be analysed.

Test was done on x20_transmission6[2018-02-27_17-55-21] image and showed more fibres being correctly found.