Edited checkBlack()
Remove the code outside checkBlack causing to loop through all pixels around the position and instead had the checkBlack() function itself look at the 3x3 array of pixels centering on the position and checking if any of those are black.

Also increased the width from 1.1*FIBREWIDTH to 1.5* so that the midpoint for fibres with odd shaped ends can be found.