Counts added
Added in counts for how many coordinates were found that weren't part of a fibre, how many fibres were found and put them into a list with how many total objects in the image there are. This gives some metric as to how well the program has found the fibres.